Added Admin users
Made it so that the first user to sign up becomes an Admin, so that they can add more users, but also made sure that non-Admins can't add or delete users, only view the user list and their profile.

While adding new users, you can decide whether or not they should be an Admin.

Also had to change the user table in the database to include the is_admin column.